requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
pydantic-settings==2.2.1
cryptography==42.0.5
python-jose[cryptography]==3.3.0
slowapi==0.1.9
redis==5.0.3
anthropic==0.25.8
jinja2==3.1.4
temporalio==1.6.0
structlog==24.1.0
sse-starlette==2.1.0
alembic==1.13.1
